fix an inaccuracy where too many yellows were given
The rules of wordle are a bit more complex than I initially realized.
When there are multiple uses of a letter in the guess, the status of
those depends on the number and locations of that letter in the answer.
Letters in the correct location (green) take precedence over present
letters in the wrong location (yellow). Letters in the wrong location
are taken in order from left to right.